requirements.txt
pyqt5~=5.15.9
opencv-python~=4.10.0.84
numpy~=1.26.4
torch~=2.3.1
ultralytics~=8.0.217
pillow~=10.4.0
matplotlib~=3.9.1